tidying up sharding code
