--- pkg/logql/sharding_mapper.go
@@ -1,5 +1,0 @@
-package logql
-
-type ShardMapper struct {
-	shards int
-}
